+++ NEWS.md
@@ -36,6 +36,7 @@ Standard library changes
   * `range` can accept the stop value as a positional argument, e.g. `range(1,10,step=2)` ([#28708]).
   * `edit` can now be called on a module to edit the file that defines it ([#29636]).
   * `diff` now supports arrays of arbitrary dimensionality and can operate over any dimension ([#29827]).
+  * `sprandn` now supports result types like `ComplexF64` or `Float32` ([#30083]).
 
 Compiler/Runtime improvements
 -----------------------------